requirements.txt
Django==4.0.3
djangorestframework==3.13.1
psycopg2==2.9.3

pytest==6.2.5
pytest-mockito==0.0.4
pytest-django==4.5.2
pytest-pythonpath==0.7.4
